BasicPythonScripts/Noise Reduction Filter/requirements.txt
librosa==0.8.1
pandas==1.3.5
numpy==1.21.5
scipy==1.9.0
matplotlib==3.6.0
pillow==7.2.1
